+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Inspect constraints;
 - New `FromText`/`ToText` to easily convert type from/to text;
 - New `inspect::Relation::kind` field;
+- New `inspect::Relation::persistence` field;
 
 ### Changed
 - `inspect::Relation::ty` is deprecated;